add sstable

+++ src/config.rs
@@ -1,3 +1,3 @@
-//! Configurations of kvlite
+//! Configurations of KVLite
 
-pub const ACTIVE_SIZE_THRESHOLD: u64 = 300;
+pub const ACTIVE_SIZE_THRESHOLD: usize = 300;
